+++ changelog.md
@@ -1,5 +1,11 @@
 # Changelog for Weavy
 
+## 8.2.1 (2021-05-19)
+
+* Fixed issue with promises on panel reset.
+* Fixed heading not being displayed correctly in notes.
+* Fixed issue with event registration.
+
 ## 8.2.0 (2021-04-16)
 
 * Added navigation history states, history event and deep link support in client.